+++ Pullenti/Ner/Definition/DefinitionAnalyzer.cs
@@ -1,6 +1,7 @@
 ﻿/*
- * Copyright (c) 2013, Pullenti. All rights reserved. Non-Commercial Freeware.
- * This class is generated using the converter UniSharping (www.unisharping.ru) from Pullenti C# project. 
+ * SDK Pullenti Lingvo, version 4.1, december 2020. Copyright (c) 2013, Pullenti. All rights reserved. 
+ * Non-Commercial Freeware and Commercial Software.
+ * This class is generated using the converter Unisharping (www.unisharping.ru) from Pullenti C# project. 
  * The latest version of the code is available on the site www.pullenti.ru
  */
 
@@ -1427,7 +1428,7 @@ namespace Pullenti.Ner.Definition
                     t = t1;
                     if (vars.Count == 0) 
                         continue;
-                    int[] inds = new int[vars.Count];
+                    int[] inds = new int[(int)vars.Count];
                     while (true) 
                     {
                         tmp.Clear();
