+++ Pullenti/Ner/Denomination/DenominationReferent.cs
@@ -1,6 +1,7 @@
 ﻿/*
- * Copyright (c) 2013, Pullenti. All rights reserved. Non-Commercial Freeware.
- * This class is generated using the converter UniSharping (www.unisharping.ru) from Pullenti C# project. 
+ * SDK Pullenti Lingvo, version 4.1, december 2020. Copyright (c) 2013, Pullenti. All rights reserved. 
+ * Non-Commercial Freeware and Commercial Software.
+ * This class is generated using the converter Unisharping (www.unisharping.ru) from Pullenti C# project. 
  * The latest version of the code is available on the site www.pullenti.ru
  */
 
@@ -138,7 +139,7 @@ namespace Pullenti.Ner.Denomination
                         ty0 = ty;
                     }
                 }
-                int[] inds = new int[items.Count];
+                int[] inds = new int[(int)items.Count];
                 for (i = 0; i < inds.Length; i++) 
                 {
                     inds[i] = 0;
