+++ Pullenti/Ner/NumberToken.cs
@@ -1,6 +1,7 @@
 ﻿/*
- * Copyright (c) 2013, Pullenti. All rights reserved. Non-Commercial Freeware.
- * This class is generated using the converter UniSharping (www.unisharping.ru) from Pullenti C# project. 
+ * SDK Pullenti Lingvo, version 4.1, december 2020. Copyright (c) 2013, Pullenti. All rights reserved. 
+ * Non-Commercial Freeware and Commercial Software.
+ * This class is generated using the converter Unisharping (www.unisharping.ru) from Pullenti C# project. 
  * The latest version of the code is available on the site www.pullenti.ru
  */
 
@@ -113,7 +114,7 @@ namespace Pullenti.Ner
             base.Deserialize(stream, kit, vers);
             if (vers == 0) 
             {
-                byte[] buf = new byte[8];
+                byte[] buf = new byte[(int)8];
                 stream.Read(buf, 0, 8);
                 long lo = BitConverter.ToInt64(buf, 0);
                 Value = lo.ToString();
